refactor: change 'product-card' for 'card'; feat: add output event to card-component
