+++ src/app/pages/products-list/card/card.component.html
@@ -1,29 +1,32 @@
-<mat-card class="card">
+<mat-card appearance="outlined" class="product-card">
     <mat-card-header>
-        <mat-card-title>{{ product.name }}</mat-card-title>
-        <mat-card-subtitle>{{ product.price }} $</mat-card-subtitle>
+        <!-- Подкатегория товара -->
+        <mat-card-subtitle>{{ product?.subCategory }}</mat-card-subtitle>
     </mat-card-header>
-    <div class="galerey">
-        <button mat-icon-button disabled>
-            <mat-icon>chevron_left</mat-icon>
-        </button>
-        <img mat-card-image alt="Product photo" [src]="product.images[0].url" />
-        <button mat-icon-button disabled>
-            <mat-icon>chevron_right</mat-icon>
-        </button>
-    </div>
-    <mat-card-content>
-        <p>Отзывы: {{ product.feedbacksCount }}</p>
-        <div class="raiting">
-            Рейтинг:
-            <mat-icon [class.star-active]="isStarActive(1)">star</mat-icon>
-            <mat-icon [class.star-active]="isStarActive(2)">star</mat-icon>
-            <mat-icon [class.star-active]="isStarActive(3)">star</mat-icon>
-            <mat-icon [class.star-active]="isStarActive(4)">star</mat-icon>
-            <mat-icon [class.star-active]="isStarActive(5)">star</mat-icon>
-        </div>
-    </mat-card-content>
-    <mat-card-actions class="mat-card-actions-align-end">
-        <button mat-button [disabled]="!product" (click)="onProductBuy($event)">Buy</button>
-    </mat-card-actions>
+    <!-- Изображение -->
+    <img mat-card-image [src]="productImage?.url" [alt]="product?.name" />
+    <mat-card-footer class="card-footer">
+        <mat-card-content>
+            <!-- Наименование товара -->
+            <mat-card-title>{{ product?.name }}</mat-card-title>
+            <p>Отзывы: {{ product?.feedbacksCount }}</p>
+            <div class="rating">
+                <mat-icon [class.star-active]="isActiveStar(1)">star</mat-icon>
+                <mat-icon [class.star-active]="isActiveStar(2)">star</mat-icon>
+                <mat-icon [class.star-active]="isActiveStar(3)">star</mat-icon>
+                <mat-icon [class.star-active]="isActiveStar(4)">star</mat-icon>
+                <mat-icon [class.star-active]="isActiveStar(5)">star</mat-icon>
+            </div>
+        </mat-card-content>
+        <mat-card-actions>
+            <button mat-button>
+                <mat-icon title="Добавить в избранное">thumb_up</mat-icon>
+                <span>В избранное</span>
+            </button>
+            <button mat-button (click)="buy($event)">
+                <mat-icon title="Добавить в корзину">shopping_cart</mat-icon>
+                <span>В корзину</span>
+            </button>
+        </mat-card-actions>
+    </mat-card-footer>
 </mat-card>
