+++ src/app/pages/products-list/card/card.component.ts
@@ -1,32 +1,32 @@
-import {ChangeDetectionStrategy, Component} from '@angular/core';
-import {MatIconModule} from '@angular/material/icon';
+import {Component, Input, output} from '@angular/core';
 import {MatButtonModule} from '@angular/material/button';
 import {MatCardModule} from '@angular/material/card';
-import {productsMock} from '../../../shared/products/products.mock';
+import {MatIconModule} from '@angular/material/icon';
+import {Product} from '../../../shared/products/product.interface';
 
+/** Компонент карточки товара */
 @Component({
-    selector: 'app-card',
+    selector: 'app-product-card',
     standalone: true,
     imports: [MatCardModule, MatButtonModule, MatIconModule],
     templateUrl: './card.component.html',
     styleUrl: './card.component.css',
-    changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class CardComponent {
-    readonly product = productsMock[0];
+    @Input() product?: Product;
 
-    onProductBuy(event: Event) {
-        event.stopPropagation();
+    readonly buyProduct = output<Event>();
 
-        if (!this.product) {
-            return;
-        }
-
-        // eslint-disable-next-line no-console
-        console.log('Buy product');
+    get productImage() {
+        return this.product?.images.at(0);
     }
 
-    isStarActive(starIndex: number): boolean {
-        return !!this.product && this.product.rating >= starIndex;
+    isActiveStar(index: number): boolean {
+        return !!this.product && this.product.rating >= index;
+    }
+
+    buy(event: Event) {
+        event.stopPropagation();
+        this.buyProduct.emit(event);
     }
 }
