--- src/app/pages/products-list/card/product-card/product-card.component.css
@@ -1,34 +1,0 @@
-:host {
-    display: flex;
-    flex: 0 1 25%;
-    margin-left: 15px;
-    margin-right: 15px;
-    margin-bottom: 15px;
-
-    .product-card {
-        width: 375px;
-    }
-
-    .card-footer {
-        margin-top: auto;
-    }
-
-    @media (max-width: 768px) {
-        flex-basis: 50%;
-    }
-
-    @media (max-width: 576px) {
-        flex-basis: 100%;
-    }
-
-    img {
-        max-width: 100%;
-        max-height: 400px;
-        margin-top: auto;
-    }
-}
-
-.star-active {
-    color: yellow;
-    text-shadow: 0 0 2px black;
-}
