+++ src/app/pages/products-list/products-list.component.ts
@@ -1,12 +1,12 @@
 import {ChangeDetectionStrategy, Component} from '@angular/core';
-import {ProductCardComponent} from './card/product-card/product-card.component';
+import {CardComponent} from './card/card.component';
 import {productsMock} from '../../shared/products/products.mock';
 
 /** Компонент списка товаров */
 @Component({
     selector: 'app-products-list',
     standalone: true,
-    imports: [ProductCardComponent],
+    imports: [CardComponent],
     templateUrl: './products-list.component.html',
     styleUrl: './products-list.component.css',
     changeDetection: ChangeDetectionStrategy.OnPush,
